+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Export `Placeholder` component, use reuse last_modified [#546](https://github.com/datagouv/udata-front/pull/546)
+- Add missing `dataset` key to `CommunityResource`type [#551](https://github.com/datagouv/udata-front/pull/551)
 
 ## 1.4.0 (2024-09-23)