+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Export `Placeholder` component, use reuse last_modified [#546](https://github.com/datagouv/udata-front/pull/546)
 - Add missing `dataset` key to `CommunityResource`type [#551](https://github.com/datagouv/udata-front/pull/551)
+- Low enphasis on last element of charts [#562](https://github.com/datagouv/udata-front/pull/562)
 
 ## 1.4.0 (2024-09-23)