+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Export `Placeholder` component, use reuse last_modified [#546](https://github.com/datagouv/udata-front/pull/546)
 - Add missing `dataset` key to `CommunityResource`type [#551](https://github.com/datagouv/udata-front/pull/551)
 - Low enphasis on last element of charts [#562](https://github.com/datagouv/udata-front/pull/562)
+- Remove markdown from dataservice card [#568](https://github.com/datagouv/udata-front/pull/568)
 
 ## 1.4.0 (2024-09-23)