+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Remove markdown from dataservice card [#568](https://github.com/datagouv/udata-front/pull/568)
 - Add Swagger UI to resources and dataservice links [#575](https://github.com/datagouv/udata-front/pull/575)
 - Add back copy button without label  [#576](https://github.com/datagouv/udata-front/pull/576)
+- Rename private to draft [#572](https://github.com/datagouv/udata-front/pull/572)
 
 ## 1.4.0 (2024-09-23)